refactor: document get 메서드 통일및 카테고리 메서드 추가 설정
