+++ src/pages/QuizRegister/QuizRegister.tsx
@@ -2,9 +2,9 @@ import { Input, InputLabel } from '@components/@common';
 import styles from './quizRegister.module.scss';
 import Button from '@components/@common/button';
 import Radio from '@components/@common/radio';
-import useGetCategory from '@hooks/fireStore/useGetCategoryList';
 import useQuizForm from '@hooks/useQuizForm';
 import { INITIAL_QUIZ } from '@constants/quiz';
+import useGetCategory from '@hooks/useGetCategory';
 
 const QuizRegister = () => {
 	const { data: category } = useGetCategory();
